% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling.R
\name{clone}
\alias{clone}
\title{Clone shape(s)}
\usage{
clone(x, n)
}
\arguments{
\item{x}{the shape to clone}

\item{n}{how many times?}
}
\description{
Nothing but \code{rep(x, n)}
}
\examples{
shapes[4] \%>\% clone(121)
shapes[14:16] \%>\% clone(27)
}
