% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cool.R
\name{as_coo}
\alias{as_coo}
\alias{is_coo}
\alias{check_coo}
\alias{as_cool}
\alias{is_cool}
\alias{check_cool}
\alias{cool}
\alias{print.cool}
\alias{plot.cool}
\title{cool class}
\usage{
as_coo(x)

is_coo(x)

check_coo(x)

as_cool(x)

is_cool(x)

check_cool(x)

cool(x)

\method{print}{cool}(x, ...)

\method{plot}{cool}(x, ...)
}
\arguments{
\item{x}{the object to build/modify or test}

\item{...}{useless, only here to comply with the generic}
}
\description{
The cool class consists in a list of matrices of class \code{coo}.
It looks like the \code{$coo} component of \code{Coo} objects in \code{Momocs}.
}
\details{
Some simple helpers and \code{print}/\code{plot} methods are available.
}
