% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross}
\alias{cross}
\title{Cross shapes}
\usage{
cross(..., ratio = 0.5, steps)
}
\arguments{
\item{...}{two shapes, or a \link{cool} or a list of shapes.
Only the first two will be crossed. In all cases, they need to be \link{homogeneize}d.}

\item{ratio}{\code{numeric} (one or more) (min=0, max=1, default=0.5): how much of the first shape?}

\item{steps}{\code{integer} if provided will use \code{ratio=seq(0, 1, length.out=steps)}}
}
\value{
a \link{cool}
}
\description{
Calculates intermediates forms between two shapes
}
\examples{
s <- shapes \%>\% pick(1:2) \%>\% homogeneize()
cross(s)
s \%>\% cross(steps=100)
s \%>\% cross(steps=20) \%>\% paper_white \%>\% draw_outlines()
}
