% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross_g}
\alias{cross_g}
\title{Cross shapes gadget}
\usage{
cross_g(..., steps = 100, interval = 1000/15)
}
\arguments{
\item{...}{two shapes, or a \link{cool} or a list of shapes.
Only the first two will be crossed. In all cases, they need to be \link{homogeneize}d.}

\item{steps}{\code{integer} to feed \link{cross}, number of intermediate steps}

\item{interval}{\code{integer} to feed \link[shiny:animationOptions]{shiny::animationOptions}, time between frames in milliseconds}
}
\value{
a \link{cool}
}
\description{
Calculates intermediates forms between two shapes using a gadget
}
\examples{
\dontrun{
s <- shapes \%>\% pick(1:2) \%>\% homogeneize()
s \%>\% cross_g
}
}
