% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momfarm-package.R
\docType{package}
\name{Momfarm}
\alias{Momfarm}
\alias{Momfarm-package}
\title{Momfarm}
\description{
The goal of Momfarm is to breed shapes,
on top of Momocs. It is a place to prototype in several directions.
}
