% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momfarm-package.R
\name{Momfarm_help}
\alias{Momfarm_help}
\alias{Momfarm_website}
\title{Momfarm helpers}
\usage{
Momfarm_help(topic = NULL)

Momfarm_website()
}
\arguments{
\item{topic}{\code{character} function name, quoted}
}
\description{
Simple utilities to smooth package use
}
