% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling.R
\name{pick}
\alias{pick}
\title{Select shape(s)}
\usage{
pick(x, i)
}
\arguments{
\item{x}{the shape to clone}

\item{i}{indices of shapes to retain}
}
\description{
Nothing but \code{x[i]}
}
\examples{
shapes[4] \%>\% clone(121)
shapes[14:16] \%>\% clone(27)
}
